sdk: Split Matrix authentication methods in a separate API
