+++ crates/matrix-sdk/src/lib.rs
@@ -38,6 +38,7 @@ pub mod config;
 mod error;
 pub mod event_handler;
 mod http_client;
+pub mod matrix_auth;
 pub mod media;
 pub mod notification_settings;
 pub mod room;
@@ -50,11 +51,7 @@ pub mod sliding_sync;
 pub mod encryption;
 
 pub use account::Account;
-#[cfg(feature = "sso-login")]
-pub use client::SsoLoginBuilder;
-pub use client::{
-    Client, ClientBuildError, ClientBuilder, LoginBuilder, LoopCtrl, SendRequest, UnknownToken,
-};
+pub use client::{Client, ClientBuildError, ClientBuilder, LoopCtrl, SendRequest, UnknownToken};
 #[cfg(feature = "image-proc")]
 pub use error::ImageError;
 pub use error::{Error, HttpError, HttpResult, RefreshTokenError, Result, RumaApiError};
