+++ crates/matrix-sdk/src/room/member.rs
@@ -49,7 +49,7 @@ impl RoomMember {
     /// # async {
     /// # let user = "example";
     /// let client = Client::new(homeserver).await.unwrap();
-    /// client.login_username(user, "password").send().await.unwrap();
+    /// client.matrix_auth().login_username(user, "password").send().await.unwrap();
     /// let room_id = room_id!("!roomid:example.com");
     /// let room = client.get_joined_room(&room_id).unwrap();
     /// let members = room.members(RoomMemberships::empty()).await.unwrap();
