+++ crates/matrix-sdk/tests/integration/client.rs
@@ -1,4 +1,4 @@
-use std::{collections::BTreeMap, str::FromStr, time::Duration};
+use std::{collections::BTreeMap, time::Duration};
 
 use assert_matches::assert_matches;
 use futures_util::FutureExt;
@@ -6,19 +6,16 @@ use matrix_sdk::{
     config::SyncSettings,
     media::{MediaFormat, MediaRequest, MediaThumbnailSize},
     sync::RoomUpdate,
-    RumaApiError, Session,
+    Session,
 };
 use matrix_sdk_test::{async_test, test_json};
 use ruma::{
     api::client::{
-        self as client_api,
-        account::register::{v3::Request as RegistrationRequest, RegistrationKind},
         directory::{
             get_public_rooms,
             get_public_rooms_filtered::{self, v3::Request as PublicRoomsFilterRequest},
         },
         media::get_content_thumbnail::v3::Method,
-        session::get_login_types::v3::LoginType,
         uiaa,
     },
     assign, device_id,
@@ -27,228 +24,12 @@ use ruma::{
     mxc_uri, room_id, uint, user_id,
 };
 use serde_json::{from_value as from_json_value, json, to_value as to_json_value};
-use url::Url;
 use wiremock::{
     matchers::{header, method, path, path_regex},
     Mock, ResponseTemplate,
 };
 
 use crate::{logged_in_client, mock_sync, no_retry_test_client};
-
-#[async_test]
-async fn login() {
-    let (client, server) = no_retry_test_client().await;
-    let homeserver = Url::from_str(&server.uri()).unwrap();
-
-    Mock::given(method("GET"))
-        .and(path("/_matrix/client/r0/login"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(&*test_json::LOGIN_TYPES))
-        .mount(&server)
-        .await;
-
-    let can_password = client
-        .get_login_types()
-        .await
-        .unwrap()
-        .flows
-        .iter()
-        .any(|flow| matches!(flow, LoginType::Password(_)));
-    assert!(can_password);
-
-    Mock::given(method("POST"))
-        .and(path("/_matrix/client/r0/login"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(&*test_json::LOGIN))
-        .mount(&server)
-        .await;
-
-    client.login_username("example", "wordpass").send().await.unwrap();
-
-    let logged_in = client.logged_in();
-    assert!(logged_in, "Client should be logged in");
-
-    assert_eq!(client.homeserver().await, homeserver);
-}
-
-#[async_test]
-async fn login_with_discovery() {
-    let (client, server) = no_retry_test_client().await;
-
-    Mock::given(method("POST"))
-        .and(path("/_matrix/client/r0/login"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(&*test_json::LOGIN_WITH_DISCOVERY))
-        .mount(&server)
-        .await;
-
-    client.login_username("example", "wordpass").send().await.unwrap();
-
-    let logged_in = client.logged_in();
-    assert!(logged_in, "Client should be logged in");
-
-    assert_eq!(client.homeserver().await.as_str(), "https://example.org/");
-}
-
-#[async_test]
-async fn login_no_discovery() {
-    let (client, server) = no_retry_test_client().await;
-
-    Mock::given(method("POST"))
-        .and(path("/_matrix/client/r0/login"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(&*test_json::LOGIN))
-        .mount(&server)
-        .await;
-
-    client.login_username("example", "wordpass").send().await.unwrap();
-
-    let logged_in = client.logged_in();
-    assert!(logged_in, "Client should be logged in");
-
-    assert_eq!(client.homeserver().await, Url::parse(&server.uri()).unwrap());
-}
-
-#[async_test]
-#[cfg(feature = "sso-login")]
-async fn login_with_sso() {
-    let (client, server) = no_retry_test_client().await;
-
-    Mock::given(method("POST"))
-        .and(path("/_matrix/client/r0/login"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(&*test_json::LOGIN))
-        .mount(&server)
-        .await;
-
-    let idp = ruma::api::client::session::get_login_types::v3::IdentityProvider::new(
-        "some-id".to_owned(),
-        "idp-name".to_owned(),
-    );
-    client
-        .login_sso(|sso_url| async move {
-            let sso_url = Url::parse(&sso_url).unwrap();
-
-            let (_, redirect) =
-                sso_url.query_pairs().find(|(key, _)| key == "redirectUrl").unwrap();
-
-            let mut redirect_url = Url::parse(&redirect).unwrap();
-            redirect_url.set_query(Some("loginToken=tinytoken"));
-
-            reqwest::get(redirect_url.to_string()).await.unwrap();
-
-            Ok(())
-        })
-        .identity_provider_id(&idp.id)
-        .await
-        .unwrap();
-
-    let logged_in = client.logged_in();
-    assert!(logged_in, "Client should be logged in");
-}
-
-#[async_test]
-async fn login_with_sso_token() {
-    let (client, server) = no_retry_test_client().await;
-
-    Mock::given(method("GET"))
-        .and(path("/_matrix/client/r0/login"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(&*test_json::LOGIN_TYPES))
-        .mount(&server)
-        .await;
-
-    let can_sso = client
-        .get_login_types()
-        .await
-        .unwrap()
-        .flows
-        .iter()
-        .any(|flow| matches!(flow, LoginType::Sso(_)));
-    assert!(can_sso);
-
-    let sso_url = client.get_sso_login_url("http://127.0.0.1:3030", None).await;
-    sso_url.unwrap();
-
-    Mock::given(method("POST"))
-        .and(path("/_matrix/client/r0/login"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(&*test_json::LOGIN))
-        .mount(&server)
-        .await;
-
-    client.login_token("averysmalltoken").send().await.unwrap();
-
-    let logged_in = client.logged_in();
-    assert!(logged_in, "Client should be logged in");
-}
-
-#[async_test]
-async fn login_error() {
-    let (client, server) = no_retry_test_client().await;
-
-    Mock::given(method("POST"))
-        .and(path("/_matrix/client/r0/login"))
-        .respond_with(ResponseTemplate::new(403).set_body_json(&*test_json::LOGIN_RESPONSE_ERR))
-        .mount(&server)
-        .await;
-
-    if let Err(err) = client.login_username("example", "wordpass").send().await {
-        if let Some(RumaApiError::ClientApi(client_api::Error { status_code, body })) =
-            err.as_ruma_api_error()
-        {
-            assert_eq!(*status_code, http::StatusCode::from_u16(403).unwrap());
-
-            if let client_api::error::ErrorBody::Standard { kind, message } = body {
-                if *kind != client_api::error::ErrorKind::Forbidden {
-                    panic!("found the wrong `ErrorKind` {kind:?}, expected `Forbidden");
-                }
-
-                assert_eq!(message, "Invalid password");
-            } else {
-                panic!("non-standard error body")
-            }
-        } else {
-            panic!("found the wrong `Error` type {err:?}, expected `Error::RumaResponse");
-        }
-    } else {
-        panic!("this request should return an `Err` variant")
-    }
-}
-
-#[async_test]
-async fn register_error() {
-    let (client, server) = no_retry_test_client().await;
-
-    Mock::given(method("POST"))
-        .and(path("/_matrix/client/r0/register"))
-        .respond_with(
-            ResponseTemplate::new(403).set_body_json(&*test_json::REGISTRATION_RESPONSE_ERR),
-        )
-        .mount(&server)
-        .await;
-
-    let user = assign!(RegistrationRequest::new(), {
-        username: Some("user".to_owned()),
-        password: Some("password".to_owned()),
-        auth: Some(uiaa::AuthData::FallbackAcknowledgement(
-            uiaa::FallbackAcknowledgement::new("foobar".to_owned()),
-        )),
-        kind: RegistrationKind::User,
-    });
-
-    if let Err(err) = client.register(user).await {
-        if let Some(client_api::Error { status_code, body }) = err.as_client_api_error() {
-            assert_eq!(*status_code, http::StatusCode::from_u16(403).unwrap());
-            if let client_api::error::ErrorBody::Standard { kind, message } = body {
-                if *kind != client_api::error::ErrorKind::Forbidden {
-                    panic!("found the wrong `ErrorKind` {kind:?}, expected `Forbidden");
-                }
-
-                assert_eq!(message, "Invalid password");
-            } else {
-                panic!("non-standard error body")
-            }
-        } else {
-            panic!("found the wrong `Error` type {err:#?}, expected `UiaaResponse`");
-        }
-    } else {
-        panic!("this request should return an `Err` variant")
-    }
-}
 
 #[async_test]
 async fn sync() {
