+++ examples/image_bot/src/main.rs
@@ -33,7 +33,11 @@ async fn login_and_sync(
     let homeserver_url = Url::parse(&homeserver_url).expect("Couldn't parse the homeserver URL");
     let client = Client::new(homeserver_url).await.unwrap();
 
-    client.login_username(&username, &password).initial_device_display_name("command bot").await?;
+    client
+        .matrix_auth()
+        .login_username(&username, &password)
+        .initial_device_display_name("command bot")
+        .await?;
 
     let response = client.sync_once(SyncSettings::default()).await.unwrap();
