+++ examples/login/src/main.rs
@@ -49,7 +49,7 @@ async fn login_and_sync(homeserver_url: String) -> anyhow::Result<()> {
 
     // First, let's figure out what login types are supported by the homeserver.
     let mut choices = Vec::new();
-    let login_types = client.get_login_types().await?.flows;
+    let login_types = client.matrix_auth().get_login_types().await?.flows;
 
     for login_type in login_types {
         match login_type {
@@ -168,6 +168,7 @@ async fn login_with_password(client: &Client) -> anyhow::Result<()> {
         password = password.trim().to_owned();
 
         match client
+            .matrix_auth()
             .login_username(&username, &password)
             .initial_device_display_name(INITIAL_DEVICE_DISPLAY_NAME)
             .await
@@ -190,7 +191,7 @@ async fn login_with_password(client: &Client) -> anyhow::Result<()> {
 async fn login_with_sso(client: &Client, idp: Option<&IdentityProvider>) -> anyhow::Result<()> {
     println!("Logging in with SSO…");
 
-    let mut login_builder = client.login_sso(|url| async move {
+    let mut login_builder = client.matrix_auth().login_sso(|url| async move {
         // Usually we would want to use a library to open the URL in the browser, but
         // let's keep it simple.
         println!("\nOpen this URL in your browser: {url}\n");
