+++ examples/persist_session/src/main.rs
@@ -124,6 +124,7 @@ async fn login(data_dir: &Path, session_file: &Path) -> anyhow::Result<Client> {
         password = password.trim().to_owned();
 
         match client
+            .matrix_auth()
             .login_username(&username, &password)
             .initial_device_display_name("persist-session client")
             .await
